Experimental x86 support
- Added experimental x86 support
   - Known issue: Stage 2 hooks fail
   - Known issue: RValue Array offset isn't found
- Fixed a number of warnings in x86
- Fixed disassembler not taking into account the current architecture
- Fixed edgecases where the YYRunnerInterface stack base would be computed improperly
- Disabled internal structure access's struct size checks in x86

+++ YYToolkit/source/YYTK/Module Interface/Interface.cpp
@@ -1138,7 +1138,7 @@ namespace YYTK
 			return AURIE_MODULE_INTERNAL_ERROR;
 
 		// Prevent ourselves from reading out-of-bounds
-		if (Index >= *m_BuiltinCount)
+		if (static_cast<int>(Index) >= *m_BuiltinCount)
 			return AURIE_INVALID_PARAMETER;
 
 		VariableInformation = &m_BuiltinArray[Index];
